Fix UT to fail on master branch and pass with this patch
